refactor: move server config to internal folder
